// frontend/src/@types/context.d.ts
export interface Entry {
  id?: string;
  title: string;
  description: string;
  created_at: Date | string;
  scheduled_at: Date | string | undefined;
}
export type EntryContextType = {
  entries: Entry[];
  saveEntry: (entry: Entry) => void;
  updateEntry: (id: string, entryData: Entry) => void;
  deleteEntry: (id: string) => void;
};